feat: add toDouble conversion to handle mixed int and double API responses

+++ lib/models/weather_model.dart
@@ -59,8 +59,8 @@ class Main {
   Main({this.temp, this.feelsLike, this.pressure, this.humidity});
 
   Main.fromJson(Map<String, dynamic> json) {
-    temp = json['temp'];
-    feelsLike = json['feels_like'];
+    temp = json['temp'].toDouble();
+    feelsLike = json['feels_like'].toDouble();
     pressure = json['pressure'];
     humidity = json['humidity'];
   }
@@ -81,7 +81,7 @@ class Wind {
   Wind({this.speed});
 
   Wind.fromJson(Map<String, dynamic> json) {
-    speed = json['speed'];
+    speed = json['speed'].toDouble();
   }
 
   Map<String, dynamic> toJson() {
